Revert "discovery: cohesion I-III, statgrab_net: remove dead plugin"
This reverts commits
- 5e7ac73ddc8c55ba37e42461a01b0033c06b9a12
- c9d38d92f81590e2ff71d3f73e7df690fed11cc2
- 5116aa0fadd7ba9dfba0454e1b3ccef325f073a8
- 5a452238684f2d3f15b63296abbed17680070bcb
